pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = "Bookstore API"
authors = ["Tiê Moreno Reis <tie.mreis@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.13"
django = "^5.0"
djangorestframework = ">=3.16.1,<4.0.0"
django-extensions = ">=4.1,<5.0"
django-debug-toolbar = ">=6.0.0,<7.0.0"
faker = ">=37.5.3,<38.0.0"
psycopg2-binary = "^2.9.10"
flake8 = "^7.3.0"
black = "^25.1.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.4.1,<9.0.0"
factory-boy = ">=3.3.3,<4.0.0"
pytest-django = ">=4.11.1,<5.0.0"



[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
